requirements.txt
Flask
flask-cors
pymongo
bcrypt
python-dotenv
